# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Pawsy social discovery API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Profile {
  id: string;
  displayName: string;
  /** @nullable */
  age?: number | null;
  animal: string;
  personality: string;
  city: string;
  interests: string[];
  lookingFor: string[];
  bio: string;
  /** @nullable */
  socialEnergy?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  isCurrentUser: boolean;
}

export interface ProfileUpdate {
  /** @minLength 1 */
  displayName?: string;
  animal?: string;
  city?: string;
  interests?: string[];
  lookingFor?: string[];
  bio?: string;
  socialEnergy?: string;
}

export interface InteractionResult {
  status: string;
  /** @nullable */
  conversationId?: string | null;
}

export interface Plan {
  id: string;
  title: string;
  activity: string;
  city: string;
  dateLabel: string;
  capacity: number;
  participantCount: number;
  creatorName: string;
  joined: boolean;
}

export interface PlanInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  activity: string;
  /** @minLength 1 */
  city: string;
  /** @minLength 1 */
  dateLabel: string;
  /**
     * @minimum 2
     * @maximum 20
     */
  capacity: number;
}

export interface Message {
  id: string;
  body: string;
  sentAt: string;
  fromCurrentUser: boolean;
}

export interface Conversation {
  id: string;
  title: string;
  participantName: string;
  participantAnimal: string;
  lastMessage: string;
  unreadCount: number;
  messages: Message[];
}

export interface MessageInput {
  /**
     * @minLength 1
     * @maxLength 2000
     */
  body: string;
}

export interface CreditPackage {
  id: string;
  /** @minimum 1 */
  credits: number;
  /** @minimum 0 */
  price: number;
  currency: string;
  label: string;
}

export interface Wallet {
  /** @minimum 0 */
  balance: number;
  packages: CreditPackage[];
}

export interface CreditCheckoutInput {
  packageId: string;
}

export interface CheckoutSession {
  status: string;
  message: string;
  /** @nullable */
  checkoutUrl?: string | null;
}

export interface Subscription {
  status: string;
  planName: string;
  /** @minimum 0 */
  monthlyPrice: number;
  currency: string;
  /** @nullable */
  nextBillingDate?: string | null;
}

export interface SubscriptionCheckoutInput {
  plan: string;
}

export type CityParameter = string;

export type AnimalParameter = string;

export type InterestParameter = string;

export type ListDiscoverProfilesParams = {
city?: CityParameter;
animal?: AnimalParameter;
interest?: InterestParameter;
};