# lib/api-zod/src/generated/types/checkoutSession.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Pawsy social discovery API
 * OpenAPI spec version: 0.1.0
 */

export interface CheckoutSession {
  status: string;
  message: string;
  /** @nullable */
  checkoutUrl?: string | null;
}