# lib/api-zod/src/generated/types/creditPackage.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Pawsy social discovery API
 * OpenAPI spec version: 0.1.0
 */

export interface CreditPackage {
  id: string;
  /** @minimum 1 */
  credits: number;
  /** @minimum 0 */
  price: number;
  currency: string;
  label: string;
}