# lib/api-zod/src/generated/types/interactionResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Pawsy social discovery API
 * OpenAPI spec version: 0.1.0
 */

export interface InteractionResult {
  status: string;
  /** @nullable */
  conversationId?: string | null;
}