# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Pawsy social discovery API
 * OpenAPI spec version: 0.1.0
 */

export interface Message {
  id: string;
  body: string;
  sentAt: Date;
  fromCurrentUser: boolean;
}