# lib/api-zod/src/generated/types/plan.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Pawsy social discovery API
 * OpenAPI spec version: 0.1.0
 */

export interface Plan {
  id: string;
  title: string;
  activity: string;
  city: string;
  dateLabel: string;
  capacity: number;
  participantCount: number;
  creatorName: string;
  joined: boolean;
}