# lib/api-zod/src/generated/types/planInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Pawsy social discovery API
 * OpenAPI spec version: 0.1.0
 */

export interface PlanInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  activity: string;
  /** @minLength 1 */
  city: string;
  /** @minLength 1 */
  dateLabel: string;
  /**
     * @minimum 2
     * @maximum 20
     */
  capacity: number;
}