# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Pawsy social discovery API
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  id: string;
  displayName: string;
  /** @nullable */
  age?: number | null;
  animal: string;
  personality: string;
  city: string;
  interests: string[];
  lookingFor: string[];
  bio: string;
  /** @nullable */
  socialEnergy?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  isCurrentUser: boolean;
}