# lib/api-zod/src/generated/types/profileUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Pawsy social discovery API
 * OpenAPI spec version: 0.1.0
 */

export interface ProfileUpdate {
  /** @minLength 1 */
  displayName?: string;
  animal?: string;
  city?: string;
  interests?: string[];
  lookingFor?: string[];
  bio?: string;
  socialEnergy?: string;
}