# lib/api-zod/src/generated/types/subscription.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Pawsy social discovery API
 * OpenAPI spec version: 0.1.0
 */

export interface Subscription {
  status: string;
  planName: string;
  /** @minimum 0 */
  monthlyPrice: number;
  currency: string;
  /** @nullable */
  nextBillingDate?: Date | null;
}