# lib/api-zod/src/generated/types/wallet.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Pawsy social discovery API
 * OpenAPI spec version: 0.1.0
 */
import type { CreditPackage } from './creditPackage';

export interface Wallet {
  /** @minimum 0 */
  balance: number;
  packages: CreditPackage[];
}